requirements.txt
requests>=2.28.2
beautifulsoup4>=4.11.1
selenium>=4.8.0
webdriver-manager>=3.8.5
pandas>=1.5.3
python-dotenv>=1.0.0
